Completed unit test, removed unnecessary conftest.py file, and restructured imports and fixtures in both e2e and unit test files.
