+++ tests/e2e/conftest.py
@@ -1,7 +1,6 @@
 import pytest
 from app import app
 
-
 @pytest.fixture(scope='module')
 def test_app():
     return app.test_client()
